trying to get this to build successfully
I've updated a handful of dependencies, and removed some others (which
have become superfluous).

Still others I've disabled temporarily to see what breaks.  Build was
failing with these enabled, but and still breaks with them
commented-out, but less severly.  And this way, I can update the
dependencies systematically.  That's the plan, anyway.

I also had to fix the way the Java source path is set, and where it is
located.  I'm not exactly sure why on that latter bit.