stop getting rrd4j the hard way
After hitting some compile-time errors caused by new features in rrd4j,
I realized that it wasn't getting the dependencies it needed.  I briefly
started looking into getting lein-javac to run Maven, when it hit me
that I was reinventing the wheel.  So I simply pulled-in rrd4j as a
project dependency from clojars, rather than using the submodule.

If this doesn't break anything else in the end, I'm ripping-out the
submodule.  Maoe was only including it as a submodule because of his
patch.