fixed name collision with clojure.core
The function "update" was being defined in "rrd4clj.io", which was
causing a name collision there and in "rrd4clj.examples", because
"update" is already defined by "clojure.core" (which is implicitly
included by everything, it seems).

I made the unsatisfying quick-fix of renaming it to "update_rrd".  There
is probably a better way, but this will have to do for now.