converting from anon+name to defn
I don't know why this needed to start its life as an anonymous function,
only to be given an name immediately after that.  Perhaps maoe
originally wrote fnk to provide keyword args because they weren't yet
available as a feature in that version of Clojure.  (But I don't know,
because I haven't dug through the changelogs to find-out.)