debugging milestone: no NullPointerException
Sweet.  I was able to narrow-down the section that's causing the
NullPointerException, by circumventing the (thin) Clojure layer of
abstraction around the Java interop code.

So far it works for defining a DataSource, and I'm about to see if it
works for an RRA definition as well.  Granted, this is not a long-term
solution, and I should figure-out why the Clojure wrapper is causing a
NullPointerException at all.

To reference the interop code directly, I needed to use import-all.