w00t!  milestone reached
I've sorted-out how to properly call a function with keyword args (and
how to properly write such a function definition).  Or at-least, I got
it to stop throwing errors, and output an actual graph (now I'll work on
getting valid data rendered).

I think this is what maoe was trying to achieve with his funky/fnk
macro.  I guess Clojure didn't have this feature, so he wrote it
himself.  When Clojure finally did implement it, the realization wasn't
far from maoe's, but it was different enough (understandably so).