removed rrd4j submodule
I've removed the rrd4j submodule and all references to it, because it
referred to an out-of-date fork of that project.  Much of the work I've
done so-far was to bring this codebase up-to-date with the latest state
of the official rrd4j (though admittedly, there is still a lot of
cleanup to do).