returned to using the function-based wrapper
I've gone back to using the function that builds the RRD object, now
that I've fixed the individual issues that were affecting components of
it.

I also changed the type definitions to use the correct syntax for Java
interop of field accessors (the syntax must have changed since this code
was last touched).