some name changes
I originally had to change update, because it was name-colliding with
something in clojure.core, and I really wasn't comfortable with that.
I've gone back and updated all of the API functions now, to be
mutually consistent.

I also renamed the RRD definition function, because the name it had was
a little too generic (potentially troublesome, in my experience).