cutting a release (v1.0.0)
This is my first attempt at a release since taking ownership of this
project.  There are probably lots of things wrong with this.  But I got
it to a minimally usable state, and I want to upload that blob of code
to clojars so I can start using it in my own projects at-least.

I'll handle issues as they arise.