[bugfix/1.0.2] fixing some minor naming issues
I had forgotten to change the namespace of the project since the
ownership transfer.  As a result, the push to Clojars was failing (after
successful auth), because I naturally wouldn't have permission to modify
Mitsutoshi's namespace.

Also added some missing key-val pairs to the project definition, because
lein-clojars was warning me about them.

Updated the license and copyright reference.